10.04.2023

+++ BankProject.sql
@@ -1,7 +1,7 @@
 CREATE TABLE client (
     id INTEGER AUTO_INCREMENT PRIMARY KEY,
-    manager_id INTEGER,
-    status INT(1),
+    manager_id smallint,
+    status tinyint,
     tax_code VARCHAR(20) unique,
     first_name VARCHAR(50),
     last_name VARCHAR(50),
@@ -17,8 +17,8 @@ CREATE TABLE account (
     id integer  AUTO_INCREMENT PRIMARY KEY,
     client_tax_code varchar(20),
     name VARCHAR(100),
-    type integer(1),
-    status integer(1),
+    type tinyint,
+    status tinyint,
     balance numeric(15,2),
     currency_code integer(2),
     created_at TIMESTAMP,
@@ -27,10 +27,10 @@ CREATE TABLE account (
 
 CREATE TABLE product (
     id INTEGER AUTO_INCREMENT PRIMARY KEY,
-    manager_id INTEGER,
+    manager_id smallint,
     product_name VARCHAR(70),
-    status INTEGER(1),
-    currency_code INTEGER(2),
+    status tinyint,
+    currency_code smallint,
     interest_rate NUMERIC(6 , 4 ),
     limit_credit NUMERIC(15 , 2 ),
     created_at TIMESTAMP,
@@ -42,7 +42,7 @@ CREATE TABLE agreement (
     account_id INTEGER,
     product_id INTEGER,
     interest_rate NUMERIC(6 , 4 ),
-    status INTEGER,
+    status tinyint,
     sum NUMERIC(15 , 2 ),
     created_at TIMESTAMP,
     updated_at TIMESTAMP
@@ -52,17 +52,17 @@ CREATE TABLE transaction (
     id INTEGER AUTO_INCREMENT PRIMARY KEY,
     debit_account INTEGER,
     credit_account INTEGER,
-    typ INTEGER(1),
+    typ tinyint,
     amount NUMERIC(15 , 2 ),
     description VARCHAR(255),
     created_at TIMESTAMP
 );
 
 CREATE TABLE manager (
-    id INTEGER AUTO_INCREMENT PRIMARY KEY,
+    id smallint AUTO_INCREMENT PRIMARY KEY,
     first_name VARCHAR(50),
     last_name VARCHAR(50),
-    status INTEGER(1),
+    status tinyint,
     description VARCHAR(255),
     created_at TIMESTAMP
 );
